build script

+++ build_lib_android.sh
@@ -1,26 +1,39 @@
+basepath=$(cd `dirname $0`; pwd)
+echo ${basepath}
+
+cd ${basepath}
+
+if [ ! -d ./Lib ];then 
+    mkdir Lib
+fi
+
 cd Lib
-rm -rf EyerLib
+rm -r EyerLib
 cd ../
 
-cd EyerLib
-cd Lib
+cd ${basepath}/EyerLib/Lib
+
 mkdir build_a
 
-rm -rf install
+if [ -d ./install ];then 
+    rm -rf install
+fi
 
 cd build_a
 
 # NDK=/Users/lichi/ndk_test/android-ndk-r21
 # NDK=/home/redknot/NDK/android-ndk-r21
 
-cmake ../ -DCMAKE_TOOLCHAIN_FILE=$NDK/build/cmake/android.toolchain.cmake -DANDROID_ABI=armeabi-v7a -DANDROID_NATIVE_API_LEVEL=21
+cmake ../ -DCMAKE_TOOLCHAIN_FILE=$NDK/build/cmake/android.toolchain.cmake -DANDROID_ABI=armeabi-v7a -DANDROID_NATIVE_API_LEVEL=21 -DBUILD_TYPE=Release
 make
 make install
 
-cd ..
-rm -rf build_a
+cd ${basepath}/EyerLib/Lib
 
-cd ..
-cd ..
+if [ -d ./build_a ];then 
+    rm -rf build_a
+fi
+
+cd ${basepath}
 
 cp -r EyerLib/Lib/install/ Lib/EyerLib
